+++ src/docs/pages/component/timepicker/index.vue
@@ -159,7 +159,7 @@
     >conf-type
     String
     >conf-default
-    `''`
+    `'请选择'`
     ---
     #demo
     >tpl